requirements.txt
llama-index-core==0.11.11
llama-index-embeddings-openai==0.2.5
llama-index-llms-openai==0.2.9
llama-index-utils-workflow==0.2.1
python-dotenv==1.0.1
gh-md-to-html[pdf_export]
google-search-results==2.4.2